refactor: applied code review comments.
