+++ product-backend/src/test/java/com/kainos/ea/service/JobRoleServiceTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.kainos.ea.api.JobRoleService;
 import org.kainos.ea.cli.JobRole;
 import org.kainos.ea.client.DatabaseConnectionException;
+import org.kainos.ea.client.FailedToGetAllJobRolesException;
 import org.kainos.ea.client.FailedToGetJobRoleException;
 import org.kainos.ea.client.JobRoleDoesNotExistException;
 import org.kainos.ea.db.DatabaseConnector;
@@ -15,10 +16,14 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 public class JobRoleServiceTest {
@@ -76,5 +81,34 @@ public class JobRoleServiceTest {
 
         assertThrows(JobRoleDoesNotExistException.class,
                 () -> jobRoleService.getJobRoleById(Id));
+    }
+
+    @Test
+    void getJobRoles_shouldReturnJobRoles_whenDaoReturnsJobRoles() throws SQLException, FailedToGetAllJobRolesException, DatabaseConnectionException {
+        List<JobRole> mockedJobRole = new ArrayList<>();
+
+        mockedJobRole.add(new JobRole(
+                1,
+                "name",
+                "description",
+                "sharePointLink"
+        ));
+
+        Mockito.when(jobRoleDao.getAllJobRoles()).thenReturn(mockedJobRole);
+
+        List<JobRole> result = jobRoleService.getAllJobRoles();
+
+        assertEquals(result, mockedJobRole);
+        verify(jobRoleDao).getAllJobRoles();
+    }
+
+    @Test
+    void getJobRole_shouldThrowsFailedToGetAllJobRolesException_whenFailedToGetAllJobs() throws SQLException, DatabaseConnectionException {
+        Mockito.when(jobRoleDao.getAllJobRoles()).thenThrow(SQLException.class);
+
+        assertThatExceptionOfType(FailedToGetAllJobRolesException.class)
+                .isThrownBy(() -> jobRoleService.getAllJobRoles())
+                .withMessage("Failed to get job all job roles");
+
     }
 }
