+++ product-web-app/tsconfig.json
@@ -28,7 +28,7 @@
 
     "moduleResolution": "Node16",
     /* Modules */
-    "module": "Node16",                                /* Specify what module code is generated. */
+    "module": "ES2020",                                /* Specify what module code is generated. */
     // "rootDir": "./",                                  /* Specify the root folder within your source files. */
     // "moduleResolution": "node10",                     /* Specify how TypeScript looks up a file from a given module specifier. */
     // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
